% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(VIStack, ROI, Percentage, Smoothing)
}
\arguments{
\item{VIStack}{- RasterStack or RasterBrick object - stack of time series vegetation index images}

\item{ROI}{-SpatialPolygon, SpatialPolygonDataFrame, SpatialPoints or SpatialPointsDataFrame object- - to define the region of interest. The ROI parameter is optional; if not provided the whole region will be considered for phenologic calculation}

\item{Percentage}{- Optional Numeric Vlaue - percentage of minimum NDVI value at which the Onset and Offset is defined. The 'Percentage' paramenter is optional; if not provided, a Default value of 20 will be taken.}

\item{Smoothing}{- Optional logical value - if the user chooses to use smoothed curve or row/unsmoothed curve. If "Smoothing' is set to TRUE, the moving average filter, with forward window size  of 2, will be applied to the vegetation index curve. The default value, if not provided, is FALSE, then the unsmoothed row data be used for the analysis.}
}
\value{
PhenoStack.img - a raster stack of 15 images in the order of OnsetV, OnsetT, MaxV, MaxT, OffsetV, OffsetT, LengthGS, BeforeMaxT, AfterMaxT, GreenUpSlope, BrownDownSlope, TINDVI, TINDVIBeforeMax, TINDVIAfterMax, Asymmetry
}
\description{
This function extracts 15 phenologic metrics from time series vegetaion index data, and provide as a RasterStack. The function takes RasterStack of vegetation index images with ooptional ROI, Percentage and Smoothing parametrs and provides a RasterStack of phenological metrics.
}
\examples{
#EXAMPLE - 1

PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE, 15, TRUE)

#EXAMPLE - 2

PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE)


}
\seealso{
MultiPointsPlot (VIStack)
}
\keyword{Phenology}
\keyword{Time-series}
\keyword{image}
\keyword{remote sensing}
\keyword{satellite}
\keyword{Crop}
\keyword{Growth stages}
