% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{SinglePhenology}
\alias{SinglePhenology}
\title{Phenology plot per pixel}
\usage{
SinglePhenology(AnnualTS, Percentage = 20, Smoothing = FALSE)
}
\arguments{
\item{Percentage}{- optional numeric value- the percentage threshold for Onset and Offset}

\item{Smoothing}{- optional boolean value- if the moving average smoothing is required or not}

\item{AnnualTS-}{annual time series}
}
\value{
return phenologic metrics for a single pixel
}
