% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{MultiPointsPlot}
\alias{MultiPointsPlot}
\title{Time series curves for Multiple points in the Region of Interest}
\usage{
MultiPointsPlot(VIStack)
}
\arguments{
\item{VIStack}{- RasterStack of time series vegetation index images}
}
\value{
Multiple time series curves together at the plot panel
}
\description{
MultiPointsPlot function allows the user to plot of vegetation dynamics curves from multiple points.
}
\details{
Plotting dynamics curves from multiple points together in a single plot helps understanding the growth variability across the field.This inforaiton allow observation of the spatial and temporal crop growth variability across the growth seasons, which provide important information about the environmental factors influencing crop growth and thus potential opportunities for influencing crop management (eg . Araya et al., 2016)
}
\seealso{
PhenoMetrics, SinglePhenology
}
\keyword{Curves}
\keyword{Dynamics}
\keyword{curves}
\keyword{multiple}
\keyword{points}
\keyword{time-series}
