% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(Path, BolAOI, Percentage, Smoothing)
}
\arguments{
\item{Path}{- Text value - the path where the time series images saved}

\item{BolAOI}{- Logical value - if there is any area of intererst or not}

\item{Percentage}{- Optional Numeric Vlaue - percentage of minimum NDVI value at which the Onset and Offset is defined. The 'Percentage' paramenter is optional; if not provided, a Default value of 10 will be taken.}

\item{Smoothing}{- Optional logical value - if the user chooses to use smoothed curve or row/unsmoothed curve. If "Smoothing' is set to TRUE, the moving avegare filter will be applied to the vegetation index curve. The default value, if not provided, is FALSE, then the unsmoothed row data be used for the analysis.}
}
\value{
OnsetV - NDVI value at the start of continuous positive slope between successive NDVI values above the user defined percentage threshold. It represent early gorth srages (seedling)

OnsetT - MODIS acquisition time when OnsetV is derived.

MaxV - the annual maximum NDVI, represents full canopy coverage during Anthesis stage

MaxT - the time when the maximum NDVI occurs, represents anthesis/flowering stage

OffsetV - NDVI value measured at the lowest slope below a user defined percentage threshold.

OffsetT - the time when the offsetV derived, represents time when crop has ripened

LengthGS- the length of the growing season between Onset and Offset

BeforeMaxT - The duration between the OnsetT and MaxT

AfterMaxT - The duration between the MaxT and OffsetT

GreenUpSlope- The rate at which NDVI increases from the OnsetV to MaxV over the time of the difference between MaxT and OnsetT

BrownDownSlope - The rate at which NDVI decreases from MaxV to OffsetV over the difference between OffsetT and MaxT

TINDVI - The sum of NDVI values attained at each image date within the growing season measured by the area under the NDVI curve.

TINDVIBeforeMax - the integral area under the curve between Onset and Maximum NDVI, indicates the pre-anthesis crop growth.

TINDVIAfterMax - the integral area under the curve between Maximum NDVI and Offset, indicates the post-anthesis growth.

Asymmetry - the difference between AreaBeforeMax and AreaAfterMax. It measures which part of the growing season attain relatively higher accumulated NDVI values
}
\description{
This function extracts 15 phenologic metrics from time series vegetaion index data, as raster and Ascii files. The function takes path of the vegetation index data and the boolean Value for BolAOI (True- if there is AOI polygon, FALSE- if the parameters are calculated for the whole region).
}
\examples{
# EXAMPLE - 1

#PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE, 15, TRUE)

# EXAMPLE - 2

#PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE)


}
\seealso{
MultiPointsPlot (Path, N,Id1, Id2...IdN)
}
\author{
Sofanit Araya
}
\keyword{Phenology,}
\keyword{Time-series}
\keyword{image,}
\keyword{remote}
\keyword{satellite}
\keyword{sensing,}
