% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(Path, BolAOI, Percentage, Smoothing)
}
\arguments{
\item{Path}{- Text value - the path where the time series images saved}

\item{BolAOI}{- Logical value - if there is any area of intererst or not}

\item{Percentage}{- Optional Numeric Vlaue - percentage of minimum NDVI value at which the Onset and Offset is defined. The 'Percentage' paramenter is optional; if not provided, a Default value of 10 will be taken.}

\item{Smoothing}{- Optional logical value - if the user chooses to use smoothed curve or row/unsmoothed curve. If "Smoothing' is set to TRUE, the moving avegare filter will be applied to the vegetation index curve. The default value, if not provided, is FALSE, then the unsmoothed row data be used for the analysis.}
}
\value{
PhenoStack.img - a raster stack of 15 images in the order of OnsetV, OnsetT, MaxV, MaxT, OffsetV, OffsetT, LengthGS, BeforeMaxT, AfterMaxT, GreenUpSlope, BrownDownSlope, TINDVI, TINDVIBeforeMax, TINDVIAfterMax, Asymmetry
}
\description{
This function extracts 15 phenologic metrics from time series vegetaion index data, as raster and Ascii files. The function takes path of the vegetation index data and the boolean Value for BolAOI (True- if there is AOI polygon, FALSE- if the parameters are calculated for the whole region).
}
\examples{
EXAMPLE - 1

PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE, 15, TRUE)

EXAMPLE - 2

PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE)


}
\seealso{
MultiPointsPlot (Path, N,Id1, Id2...IdN)
}
\author{
Sofanit Araya, Bertram Ostendorf, Megan Lewis and Greg Lyle
}
\keyword{Phenology,}
\keyword{Time-series}
\keyword{image,}
\keyword{remote}
\keyword{satellite}
\keyword{sensing,}
