% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoPlot}
\alias{PhenoPlot}
\title{Plot phenological metrics in a raster format Phenology plot per pixel}
\usage{
PhenoPlot(PhenoStack)
}
\arguments{
\item{PhenoStack}{- RasterStack object of phenological metrics}
}
\value{
Plots  15 phenological metrics
}
\description{
PhenoPlot takes the output file of the PhenoMetrics function and plot the raster matrics.
}
\seealso{
MultiPointsPlot, PhenoMetrics, SinglePhenology
}
