% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{SinglePhenology}
\alias{SinglePhenology}
\title{Phenology plot per pixel}
\usage{
SinglePhenology(AnnualTS, Percentage = NULL, Smoothing = FALSE)
}
\arguments{
\item{AnnualTS}{- annual time series}

\item{Percentage}{- the percentage threshold for Onset and Offset}

\item{Smoothing}{- moving average smoothing applied if TRUE}
}
\value{
return phenologic metrics for a single pixel as an array in the sequence of: OnsetV, OnsetT, OffsetV, OffsetT, MAxV, MaxT, TINDVI, TINDVIBeforeMax, TINDVIAfeterMax, Assymetry, GreenUpSlope, BrownDownSlope, LengthGS, BeforeMaxT, AfterMaxT".
}
\description{
SinglePhenology is a premitive function which takes a time series vegetation index data for a single pixel for a single season.
}
\seealso{
MultiPointsPlot, PhenoMetrics
}
