% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{SinglePhenology}
\alias{SinglePhenology}
\title{Phenology plot per pixel}
\usage{
SinglePhenology(AnnualTS, Percentage = NULL, Smoothing = FALSE)
}
\arguments{
\item{AnnualTS}{- annual time series}

\item{Percentage}{- the percentage threshold for Onset and Offset}

\item{Smoothing}{- moving average smoothing applied if TRUE}
}
\value{
return phenologic metrics for a single pixel as an array in the sequence of: OnsetV, OnsetT, OffsetV, OffsetT, MAxV, MaxT, TINDVI, TINDVIBeforeMax, TINDVIAfeterMax, Assymetry, GreenUpSlope, BrownDownSlope, LengthGS, BeforeMaxT, AfterMaxT".
}
\description{
SinglePhenology is a premitive function which takes a time series vegetation index data for a single pixel for a single season.
}
\seealso{
MultiPointsPlot, PhenoMetrics

#Example
Point <- c(0.2052, 0.1824, 0.1780, 0.1732, 0.1861, 0.1863, 0.1884, 0.2202, 0.2669, 0.2708, 0.3700, 0.5900, 0.6909, 0.6057, 0.6750, 0.5572, 0.4990, 0.3463, 0.2579, 0.2167, 0.1672, 0.1771, 0.1856)
SinglePhenology(Point)
}
